requirements.txt
altair
annotated-doc
annotated-types
anyio
attrs
bcrypt
blinker
build
cachetools
certifi
cffi
charset-normalizer
chromadb
click
cryptography
durationpy
et_xmlfile
fastapi
filelock
flatbuffers
fsspec
gitdb
GitPython
googleapis-common-protos
grpcio
h11
hf-xet
httpcore
httptools
httpx
huggingface_hub
idna
importlib_metadata
importlib_resources
jieba
Jinja2
jsonschema
jsonschema-specifications
kubernetes
litellm==1.82.0
lxml
markdown-it-py
MarkupSafe
mdurl
mmh3
mpmath
narwhals
numpy
oauthlib
onnxruntime
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
pandas
opendataloader-pdf[hybrid]
paddleocr
pillow
protobuf
pyarrow
pybase64
pycparser
pydantic
pydantic-settings
pydantic_core
pydeck
Pygments
pymupdf==1.26.4
PyPDF2==3.0.1
pypdfium2
PyPika
pyproject_hooks

python-dateutil
python-docx
python-dotenv==1.1.0
pytz
PyYAML
rank-bm25
referencing
requests
requests-oauthlib
rich
rpds-py
shellingham
six
smmap
starlette
streamlit
sympy
tenacity
tokenizers
toml
tornado
tqdm
typer
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
uvloop
watchdog
watchfiles
websocket-client
websockets
zipp
